Fixed remaining compile warnings.
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@9477 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ renderer/VertexCache.cpp
@@ -242,8 +242,10 @@ void idVertexCache::EndFrame() {
 	indexAllocCount = indexUseCount = vertexAllocCount = vertexUseCount = 0;
 
 	// check if we need to resize current buffer set
-	if ( dynamicData.vertexBuffer.BytesRemaining() < currentVertexCacheSize
-			|| dynamicData.indexBuffer.BytesRemaining() < currentIndexCacheSize ) {
+	if (
+		(int)dynamicData.vertexBuffer.BytesRemaining() < currentVertexCacheSize ||
+		(int)dynamicData.indexBuffer.BytesRemaining() < currentIndexCacheSize
+	) {
 		common->Printf( "Resizing dynamic VertexCache: index %d kb -> %d kb, vertex %d kb -> %d kb\n", dynamicData.indexBuffer.BytesRemaining() / 1024, currentIndexCacheSize / 1024, dynamicData.indexBuffer.BytesRemaining() / 1024, currentVertexCacheSize / 1024 );
 		FreeGeoBufferSet( dynamicData );
 		AllocGeoBufferSet( dynamicData, currentVertexCacheSize, currentIndexCacheSize );
